% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_data.R
\name{clean_and_geocode}
\alias{clean_and_geocode}
\title{Clean and geocode the data}
\usage{
clean_and_geocode()
}
\value{
Writes a CSV file named "Data_geo.csv"
}
\description{
This function imports the raw dataset, cleans it, and uses OSM API (nominatim) to geocode (i.e detect the latitude and longitude) the cities.
}
