% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conversion_factors}
\alias{conversion_factors}
\title{Conversion of distance in a carbon footprint}
\format{
A data frame with 20 rows and 10 variables
}
\description{
A dataset containing the level of CO2 (among others) by categories of
distance (short, medium, long, international) and of flight class (unknown, economy, economy+, business, first).
}
\details{
This data comes from the package footprint (\code{footprint:::conversion_factors}).
}
\keyword{datasets}
