% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_distance.R
\name{custom_dist}
\alias{custom_dist}
\title{Function to compute the distance between two points}
\usage{
custom_dist(departure_lat, departure_long, arrival_lat, arrival_long)
}
\arguments{
\item{departure_lat}{Latitude of the first city}

\item{departure_long}{Longitude of the first city}

\item{arrival_lat}{Latitude of the second city}

\item{arrival_long}{Longitude of the second city}
}
\value{
A numeric value
}
\description{
This function was adapted from the source code of footprint::latlong_footprint(). Credits go to Anthony Schmidt.
}
