% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_footprint.R
\name{custom_footprint}
\alias{custom_footprint}
\title{Title}
\usage{
custom_footprint(
  departure_lat,
  departure_long,
  arrival_lat,
  arrival_long,
  flightClass = "Unknown",
  output = "co2e"
)
}
\arguments{
\item{departure_lat}{}

\item{departure_long}{}

\item{arrival_lat}{}

\item{arrival_long}{}

\item{flightClass}{}

\item{output}{}
}
\description{
Title
}
