% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_distance.R
\name{dist_player_year}
\alias{dist_player_year}
\title{Compute the total distance for a specific player and year}
\usage{
dist_player_year(player, year)
}
\arguments{
\item{player}{Name of the player}

\item{year}{Year}
}
\value{
A number (in km)
}
\description{
Compute the total distance for a specific player and year
}
