% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_distance.R
\name{evol_dist}
\alias{evol_dist}
\title{Compute the evolution of distance per year for a specific player}
\usage{
evol_dist(player)
}
\arguments{
\item{player}{Player name}
}
\value{
A tibble
}
\description{
Compute the evolution of distance per year for a specific player
}
