% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_data.R
\name{filter_player_year}
\alias{filter_player_year}
\title{Filter by player and year}
\usage{
filter_player_year(player, year)
}
\arguments{
\item{player}{Player name}

\item{year}{Year}
}
\value{
A tibble
}
\description{
Format correctly latitudes and longitudes to compute distance
}
