% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_footprint.R
\name{footprint_player_year}
\alias{footprint_player_year}
\title{Compute the total footprint for a specific player and year}
\usage{
footprint_player_year(player, year, output = "co2e")
}
\arguments{
\item{player}{Player name}

\item{year}{Year}

\item{output}{Output type, can take several values, see ?footprint::latlong_footprint}
}
\value{
A number (in kilograms of output)
}
\description{
Compute the total footprint for a specific player and year
}
