% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_data.R
\name{get_tennis_data}
\alias{get_tennis_data}
\title{Get the data and create the order in which the players play the tournaments}
\usage{
get_tennis_data()
}
\value{
A tibble
}
\description{
Get the data and create the order in which the players play the tournaments
}
