% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{map_travel}
\alias{map_travel}
\title{Show on a map the trips of a particular player for a a year}
\usage{
map_travel(player, year)
}
\arguments{
\item{player}{Player name}

\item{year}{Year}
}
\description{
Show on a map the trips of a particular player for a a year
}
