% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evol.R
\name{plot_evol}
\alias{plot_evol}
\title{Plot the footprint or the distance made by one or several players}
\usage{
plot_evol(player, indicator)
}
\arguments{
\item{player}{Player(s) name(s)}

\item{indicator}{"dist" or "footprint"}
}
\value{
A graph
}
\description{
Plot the footprint or the distance made by one or several players
}
