% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis_data}
\alias{tennis_data}
\title{Final dataset}
\format{
A data frame with 164804 rows and 21 variables
}
\description{
A dataset combining \code{tennis_players} and \code{tennis_data_geocodes}.
}
\keyword{datasets}
