% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis_data_geocodes}
\alias{tennis_data_geocodes}
\title{List of pairs player-tournament}
\format{
A data frame with 166065 rows and 9 variables
}
\description{
A dataset containing every pair player-tournament for each year. It also contains the latitude and longitude of the tournament.
}
\details{
This data comes from https://github.com/JeffSackmann/tennis_atp.
}
\keyword{datasets}
