% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis_players}
\alias{tennis_players}
\title{List of male tennis players}
\format{
A data frame with 40957 rows and 7 variables
}
\description{
A dataset containing the name, date of birth, and country of origin of all male tennis players since 1968.
}
\details{
This data comes from https://github.com/JeffSackmann/tennis_atp/blob/master/atp_players.csv.
}
\keyword{datasets}
