% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{tournament_calendar}
\alias{tournament_calendar}
\title{Show the location of tournaments month per month for a given year}
\usage{
tournament_calendar(year)
}
\arguments{
\item{year}{Year}
}
\value{
A plot
}
\description{
Show the location of tournaments month per month for a given year
}
