% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_footprint.R
\name{custom_footprint}
\alias{custom_footprint}
\title{Title}
\usage{
custom_footprint(
  departure_lat,
  departure_long,
  arrival_lat,
  arrival_long,
  flightClass = "Unknown",
  output = "co2e"
)
}
\arguments{
\item{departure_lat}{Latitude of departure city}

\item{departure_long}{Longitude of departure city}

\item{arrival_lat}{Latitude of arrival city}

\item{arrival_long}{Longitude of arrival city}

\item{flightClass}{Flight class of individual}

\item{output}{Unit of carbon footprint}
}
\value{

}
\description{
Title
}
